+++ src/test/java/org/panteleyev/persistence/UuidPrimaryKeyTest.java
@@ -28,12 +28,15 @@ package org.panteleyev.persistence;
 
 import org.panteleyev.persistence.base.Base;
 import org.panteleyev.persistence.model.UuidPrimaryKeyRecord;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 import java.util.Collections;
 import java.util.UUID;
 import static org.panteleyev.persistence.base.Base.MYSQL_GROUP;
 import static org.panteleyev.persistence.base.Base.SQLITE_GROUP;
 import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertNull;
+import static org.testng.Assert.assertTrue;
 
 @Test(groups = {SQLITE_GROUP, MYSQL_GROUP})
 public class UuidPrimaryKeyTest extends Base {
@@ -48,14 +51,49 @@ public class UuidPrimaryKeyTest extends Base {
 
         getDao().insert(record);
 
-        var retrieved = getDao().get(id, UuidPrimaryKeyRecord.class);
-        assertEquals(retrieved, record);
+        getDao().get(id, UuidPrimaryKeyRecord.class)
+            .ifPresentOrElse(retrieved -> assertEquals(retrieved, record), Assert::fail);
 
         var newValue = UUID.randomUUID().toString();
         getDao().update(new UuidPrimaryKeyRecord(record.getPrimKey(), newValue));
 
-        var updated = getDao().get(id, UuidPrimaryKeyRecord.class);
-        assertEquals(updated.getPrimKey(), record.getPrimKey());
-        assertEquals(updated.getValue(), newValue);
+        getDao().get(id, UuidPrimaryKeyRecord.class).ifPresentOrElse(updated -> {
+            assertEquals(updated.getPrimKey(), record.getPrimKey());
+            assertEquals(updated.getValue(), newValue);
+        }, Assert::fail);
+    }
+
+    @Test
+    public void testUuidDeleteByRecord() {
+        getDao().createTables(Collections.singletonList(UuidPrimaryKeyRecord.class));
+        getDao().preload(Collections.singletonList(UuidPrimaryKeyRecord.class));
+
+        var id = UUID.randomUUID();
+        var record = new UuidPrimaryKeyRecord(id, UUID.randomUUID().toString());
+
+        getDao().insert(record);
+        getDao().get(id, UuidPrimaryKeyRecord.class)
+            .ifPresentOrElse(retrieved -> assertEquals(retrieved, record), Assert::fail);
+
+        getDao().delete(record);
+        var deleted = getDao().get(id, UuidPrimaryKeyRecord.class);
+        assertTrue(deleted.isEmpty());
+    }
+
+    @Test
+    public void testUuidDeleteById() {
+        getDao().createTables(Collections.singletonList(UuidPrimaryKeyRecord.class));
+        getDao().preload(Collections.singletonList(UuidPrimaryKeyRecord.class));
+
+        var id = UUID.randomUUID();
+        var record = new UuidPrimaryKeyRecord(id, UUID.randomUUID().toString());
+
+        getDao().insert(record);
+        getDao().get(id, UuidPrimaryKeyRecord.class)
+            .ifPresentOrElse(retrieved -> assertEquals(retrieved, record), Assert::fail);
+
+        getDao().delete(id, UuidPrimaryKeyRecord.class);
+        var deleted = getDao().get(id, UuidPrimaryKeyRecord.class);
+        assertTrue(deleted.isEmpty());
     }
 }