+++ src/test/java/org/panteleyev/persistence/base/DatabaseType.java
@@ -1,0 +1,6 @@
+package org.panteleyev.persistence.base;
+
+public enum DatabaseType {
+    SQLITE,
+    MYSQL
+}