+++ src/test/java/org/panteleyev/persistence/model/SelfReferencingTable.java
@@ -34,44 +34,60 @@ import org.panteleyev.persistence.annotations.RecordBuilder;
 import org.panteleyev.persistence.annotations.ReferenceOption;
 import org.panteleyev.persistence.annotations.Table;
 
-@Table("child_table")
-public class ChildTable implements Record<Integer> {
+@Table("self_referencing_table")
+public class SelfReferencingTable implements Record<Integer> {
     @PrimaryKey
     @Column(Column.ID)
     private int id;
 
-    @Column("null_value")
-    @ForeignKey(table = ParentTable.class, column = "value",
+    @Column(value = "value", unique = true)
+    private String value;
+
+    @Column(value = "null_value")
+    @ForeignKey(table = SelfReferencingTable.class, column = "value",
             onDelete = ReferenceOption.SET_NULL, onUpdate = ReferenceOption.SET_NULL)
     private final String nullValue;
 
-    @Column("cascade_value")
-    @ForeignKey(table = ParentTable.class, column = "value",
+    @Column(value = "cascade_value")
+    @ForeignKey(table = SelfReferencingTable.class, column = "value",
             onDelete = ReferenceOption.CASCADE, onUpdate = ReferenceOption.CASCADE)
     private final String cascadeValue;
 
     @Column("restrict_value")
-    @ForeignKey(table = ParentTable.class, column = "value",
+    @ForeignKey(table = SelfReferencingTable.class, column = "value",
             onDelete = ReferenceOption.RESTRICT, onUpdate = ReferenceOption.RESTRICT)
     private final String restrictValue;
 
     @Column("no_action_value")
-    @ForeignKey(table = ParentTable.class, column = "value",
+    @ForeignKey(table = SelfReferencingTable.class, column = "value",
             onDelete = ReferenceOption.NO_ACTION, onUpdate = ReferenceOption.NO_ACTION)
     private final String noActionValue;
 
     @Column("none_value")
-    @ForeignKey(table = ParentTable.class, column = "value")
+    @ForeignKey(table = SelfReferencingTable.class, column = "value")
     private final String noneValue;
 
-    @RecordBuilder
-    public ChildTable(@Column("id") int id,
-                      @Column("null_value") String nullValue,
-                      @Column("cascade_value") String cascadeValue,
-                      @Column("restrict_value") String restrictValue,
-                      @Column("no_action_value") String noActionValue,
-                      @Column("none_value") String noneValue) {
+    public SelfReferencingTable(int id, String value) {
         this.id = id;
+        this.value = value;
+
+        this.nullValue = null;
+        this.cascadeValue = null;
+        this.restrictValue = null;
+        this.noActionValue = null;
+        this.noneValue = null;
+    }
+
+    @RecordBuilder
+    public SelfReferencingTable(@Column("id") int id,
+                                @Column("value") String value,
+                                @Column("null_value") String nullValue,
+                                @Column("cascade_value") String cascadeValue,
+                                @Column("restrict_value") String restrictValue,
+                                @Column("no_action_value") String noActionValue,
+                                @Column("none_value") String noneValue) {
+        this.id = id;
+        this.value = value;
         this.nullValue = nullValue;
         this.cascadeValue = cascadeValue;
         this.restrictValue = restrictValue;
@@ -81,6 +97,14 @@ public class ChildTable implements Record<Integer> {
 
     public int getId() {
         return id;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
     }
 
     public String getNullValue() {